drm_hwcomposer: Correct CTM conversion
Correctly compute the CTM given the HAL interface.

HAL provides a 4x4 float type matrix:
| 0  1  2  3|
| 4  5  6  7|
| 8  9 10 11|
|12 13 14 15|

This is to be interpreted as
out   matrix     in    offset
|R|   |0 4  8|   |R|   |12|
|G| = |1 5  9| x |G| + |13|
|B|   |2 6 10|   |B|   |14|

DRM expects a 3x3 s31.32 fixed point matrix interpreted as
|R|   |0 1 2|   |R|
|G| = |3 4 5| x |G|
|B|   |6 7 8|   |B|

DRM does not support an offset vector, so we reject any color transform
matrices that have one.

Change-Id: I323fde281c8a97cf319bb9774eb232abcc8bde34
Signed-off-by: Sasha McIntosh <sashamcintosh@google.com>